chore: move modal core components to the UI package
