+++ web/core/components/api-token/delete-token-modal.tsx
@@ -6,9 +6,7 @@ import { mutate } from "swr";
 // types
 import { IApiToken } from "@plane/types";
 // ui
-import { TOAST_TYPE, setToast } from "@plane/ui";
-// components
-import { AlertModalCore } from "@/components/core";
+import { AlertModalCore, TOAST_TYPE, setToast } from "@plane/ui";
 // fetch-keys
 import { API_TOKENS_LIST } from "@/constants/fetch-keys";
 // services
