+++ web/core/components/core/modals/index.ts
@@ -1,9 +1,7 @@
-export * from "./alert-modal";
 export * from "./bulk-delete-issues-modal";
 export * from "./existing-issues-list-modal";
 export * from "./gpt-assistant-popover";
 export * from "./link-modal";
-export * from "./modal-core";
 export * from "./user-image-upload-modal";
 export * from "./workspace-image-upload-modal";
 export * from "./issue-search-modal-empty-state";
