+++ web/core/components/cycles/delete-modal.tsx
@@ -6,9 +6,7 @@ import { useParams, useRouter, useSearchParams } from "next/navigation";
 // types
 import { ICycle } from "@plane/types";
 // ui
-import { TOAST_TYPE, setToast } from "@plane/ui";
-// components
-import { AlertModalCore } from "@/components/core";
+import { AlertModalCore, TOAST_TYPE, setToast } from "@plane/ui";
 // constants
 import { CYCLE_DELETED } from "@/constants/event-tracker";
 // hooks
