+++ web/core/components/estimates/create/modal.tsx
@@ -3,10 +3,11 @@
 import { FC, useEffect, useMemo, useState } from "react";
 import { observer } from "mobx-react";
 import { ChevronLeft } from "lucide-react";
+// types
 import { IEstimateFormData, TEstimateSystemKeys, TEstimatePointsObject } from "@plane/types";
-import { Button, TOAST_TYPE, setToast } from "@plane/ui";
+// ui
+import { Button, EModalPosition, EModalWidth, ModalCore, TOAST_TYPE, setToast } from "@plane/ui";
 // components
-import { EModalPosition, EModalWidth, ModalCore } from "@/components/core";
 import { EstimateCreateStageOne, EstimatePointCreateRoot } from "@/components/estimates";
 // constants
 import { EEstimateSystem, ESTIMATE_SYSTEMS } from "@/constants/estimates";
