+++ web/core/components/inbox/modals/decline-issue-modal.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
 // types
 import type { TIssue } from "@plane/types";
-// components
-import { AlertModalCore } from "@/components/core";
+// ui
+import { AlertModalCore } from "@plane/ui";
 // hooks
 import { useProject } from "@/hooks/store";
